Update css to make links cover whole square
This updates the css with a new selector that targets the links inside of the blocks. This expands them using positive padding and negative margin so that they don't influence the layout, but that the link target now covers the whole square. In addition we use relative position and zindex of 1 to make the links go above the keyword text.