Make entire div a link
This helps with usability because none of the other content of the
boxes leads anywhere else.

We accomplish this thanks to this helpful stackoverflow solution: https://stackoverflow.com/a/3494108

Briefly, we create an empty span within the link, make the span take up
the entire div and make the position of the div relative.