Add name block attribute and styling
This adds a "name block" attribute that allows the block to be
interpreted as a block containing my name. This allows me to style this
block differently so that the text doesn't look crappy.

More stuff will be added to this name block, including some associated
JS and CSS that means the card will flip when clicked to reveal a very
brief bio.