Fix support for webp and add jxr support
This fixes the webp detection and adds jpeg xr detection.

Previously the webp support detection was failing because modernizr
returns some sort of weird franken-boolean as a result which actually
contains the support for webp animation, webp alpha and webp lossless.
Since all I actually need to know is if webp alpha is supported, I
change to just detecting that which solves the fact that webps were
still being served on Edge.

This also adds jpegxr as a fallback if webp is not supported.